pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "planmod"
version = "0.1.0"
description = "Cloud-based pipeline for converting drawings to structured DXF files with component-level reasoning"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [
    { name = "PlanMod Team" }
]
keywords = ["dxf", "cad", "drawing", "vectorization", "ai", "vlm", "serverless"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # AWS SDK
    "boto3>=1.34.0",
    "botocore>=1.34.0",
    
    # Image Processing
    "opencv-python-headless>=4.9.0",
    "Pillow>=10.2.0",
    "numpy>=1.26.0",
    
    # PDF Processing
    "pdf2image>=1.17.0",
    "PyMuPDF>=1.23.0",
    
    # DXF/DWG Processing
    "ezdxf>=1.1.0",
    
    # OCR
    "pytesseract>=0.3.10",
    
    # Data Models
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    
    # Configuration
    "PyYAML>=6.0",
    "python-dotenv>=1.0.0",
    
    # HTTP/API
    "httpx>=0.26.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "python-multipart>=0.0.6",
    
    # Async
    "aiofiles>=23.2.0",
    
    # Utilities
    "structlog>=24.1.0",
    "tenacity>=8.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "black>=24.1.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
    "ipython>=8.20.0",
]

cdk = [
    "aws-cdk-lib>=2.120.0",
    "constructs>=10.3.0",
]

[project.scripts]
planmod = "backend.cli:main"
planmod-api = "backend.api.server:main"

[project.urls]
Homepage = "https://github.com/planmod/planmod"
Documentation = "https://planmod.readthedocs.io"
Repository = "https://github.com/planmod/planmod"

[tool.hatch.build.targets.wheel]
packages = ["backend"]

[tool.black]
line-length = 100
target-version = ["py311"]
exclude = '''
/(
    \.git
    | \.hatch
    | \.mypy_cache
    | \.venv
    | _build
    | build
    | dist
    | node_modules
)/
'''

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.isort]
known-first-party = ["backend"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=backend --cov-report=term-missing"
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["backend"]
omit = ["tests/*", "*/__init__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]


